move tests into folder
