+++ src/datasource.test.ts
@@ -1,14 +1,14 @@
 import moment from 'moment';
 
-import TemplateSrv from './lib/template_srv_stub';
+import TemplateSrv from './mocks/template_srv_stub';
 
-import Datasource from '../src/datasource';
+import Datasource from './datasource';
 
 describe('InfluxDB (Flux)', () => {
   const templateSrv = new TemplateSrv();
-  const ds = new Datasource({url: ''}, {}, templateSrv);
+  const ds = new Datasource({ url: '' }, {}, templateSrv);
   const DEFAULT_OPTIONS = {
-    rangeRaw: {to: 'now', from: 'now - 3h'},
+    rangeRaw: { to: 'now', from: 'now - 3h' },
     scopedVars: {},
     targets: [],
   };
@@ -18,10 +18,7 @@ describe('InfluxDB (Flux)', () => {
     let target: any;
 
     it.skip('replaces $range variable', () => {
-      target = ds.prepareQueryTarget(
-        {query: 'from(bucket: "test") |> range($range)'},
-        DEFAULT_OPTIONS
-      );
+      target = ds.prepareQueryTarget({ query: 'from(bucket: "test") |> range($range)' }, DEFAULT_OPTIONS);
       expect(target.query).toBe('from(bucket: "test") |> range(start: -3h)');
     });
 
@@ -29,17 +26,15 @@ describe('InfluxDB (Flux)', () => {
       const to = moment();
       const from = moment().subtract(1, 'hours');
       target = ds.prepareQueryTarget(
-        {query: 'from(bucket: "test") |> range($range)'},
+        { query: 'from(bucket: "test") |> range($range)' },
         {
           ...DEFAULT_OPTIONS,
-          rangeRaw: {to, from},
+          rangeRaw: { to, from },
         }
       );
       const start = from.toISOString();
       const stop = to.toISOString();
-      expect(target.query).toBe(
-        `from(bucket: "test") |> range(start: ${start}, stop: ${stop})`
-      );
+      expect(target.query).toBe(`from(bucket: "test") |> range(start: ${start}, stop: ${stop})`);
     });
   });
 });
